Refactor over to a trait for the model instead
