+++ app/Http/Controllers/Assets/AssetFilesController.php
@@ -4,13 +4,15 @@ namespace App\Http\Controllers\Assets;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\AssetFileRequest;
+use App\Http\Traits\Attachable;
 use App\Models\Actionlog;
 use App\Models\Asset;
-use App\Services\FileStorageService;
 use Illuminate\Http\RedirectResponse;
 
 class AssetFilesController extends Controller
 {
+    use Attachable;
+
     /**
      * Upload a file to the server.
      *
@@ -28,7 +30,7 @@ class AssetFilesController extends Controller
         }
 
         $this->authorize('update', $asset);
-        FileStorageService::store('assets', $request->file('file'), $asset, $request->get('notes'), 'hardware');
+        $this->storeFile('assets', $request->file('file'), $asset, $request->get('notes'), 'hardware');
         return redirect()->back()->with('success', trans('admin/hardware/message.upload.success'));
     }
 
@@ -55,7 +57,7 @@ class AssetFilesController extends Controller
             return response('No matching record for that asset/file', 500)
                 ->header('Content-Type', 'text/plain');
         }
-        return FileStorageService::show($log->action_type == 'audit' ? 'audits' : 'assets', $log->filename);
+        return $this->showFile($log->action_type == 'audit' ? 'audits' : 'assets', $log->filename);
     }
 
     /**
@@ -77,7 +79,7 @@ class AssetFilesController extends Controller
 
         $this->authorize('update', $asset);
         if ($log = Actionlog::find($fileId)) {
-            FileStorageService::destroy('assets', $log);
+            $this->destroyFile('assets', $log);
         }
         return redirect()->back()->with('success', trans('admin/hardware/message.deletefile.success'));
     }
