+++ app/Http/Traits/Attachable.php
@@ -1,40 +1,21 @@
 <?php
 
-namespace App\Services;
+namespace App\Http\Traits;
 
 use App\Helpers\StorageHelper;
 use App\Models\Actionlog;
 use enshrined\svgSanitize\Sanitizer;
-use Exception;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Storage;
 
-class FileStorageService
+trait Attachable
 {
-
-    private static string $uploadsDir = 'private_uploads/';
-
-    public static function generateFileName($slug, $id, UploadedFile $file): string
-    {
-        return
-            $slug .
-            '-' .
-            $id .
-            '-' .
-            str_random(8) .
-            '-' .
-            str_slug(basename($file->getClientOriginalName(), '.' . $file->getClientOriginalExtension())) .
-            '.' .
-            $file->getClientOriginalExtension();
-    }
-
-    public static function store(string $directory, array $files, Model $model, $notes, string $fileSlug = ''): void {
+    public function storeFile(string $directory, array $files, Model $model, $notes, string $fileSlug = ''): void {
         $fileSlug = $fileSlug ?? str_singular($directory);
 
         foreach ($files as $file) {
-            $fileName = self::generateFileName($fileSlug, $model->id, $file);
+            $fileName = StorageHelper::generateFileName($fileSlug, $model->id, $file);
 
             // Check for SVG and sanitize it
             if ($file->getClientOriginalExtension() === 'svg') {
@@ -42,7 +23,7 @@ class FileStorageService
                 $sanitizer = new Sanitizer();
                 try {
                     Storage::put(
-                        self::$uploadsDir . $directory . '/' . $fileName,
+                        StorageHelper::uploadsDirectory . $directory . '/' . $fileName,
                         $sanitizer->sanitize($file->getContent())
                     );
                 } catch (Exception $e) {
@@ -50,22 +31,22 @@ class FileStorageService
                     Log::debug($e);
                 }
             } else {
-                Storage::put(self::$uploadsDir . $directory . '/' . $fileName, $file->getContent());
+                Storage::put(StorageHelper::uploadsDirectory . $directory . '/' . $fileName, $file->getContent());
             }
             $model->logUpload($fileName, e($notes));
         }
     }
 
-    public static function destroy(string $directory, ActionLog $log): void
+    public function destroyFile(string $directory, ActionLog $log): void
     {
         if ($log->delete()) {
-            Storage::delete(self::$uploadsDir . $directory . '/' . $log->filename);
+            Storage::delete(StorageHelper::uploadsDirectory . $directory . '/' . $log->filename);
         }
     }
 
-    public static function show(string $directory, string $fileName)
+    public function showFile(string $directory, string $fileName)
     {
-        $file = self::$uploadsDir . '/' . $directory . '/' . $fileName;
+        $file = StorageHelper::uploadsDirectory . '/' . $directory . '/' . $fileName;
         Log::debug('Checking for ' . $file);
         if (Storage::missing($file)) {
             return response('File ' . $file . ' not found on server', 404)
